exfatprogs: support creating the root inode in exfat_alloc_exfat()
If the root directory is corrupted, fsck needs to try to repair it
instead of returning an error code. Therefore, fsck needs to create
and initialize the root inode itself, passing a non-NULL root inode
to exfat_alloc_exfat().

For other utils, if the root directory is corrupted, they only need
to return an error code. Therefore, passing a NULL root inode to
exfat_alloc_exfat() and let exfat_alloc_exfat() be responsible for
creating and initializing the root inode.

Signed-off-by: Yuezhang Mo <Yuezhang.Mo@sony.com>
Reviewed-by: Hyunchul Lee <hyc.lee@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>