dump: fix dumping of volume-label/bitmap/upcase entries
The exfat specification only requires these three entries to be in
the root directory, but does not require them to be the first three
entries in the root directory. Therefore, it cannot be assumed that
the first three entries in the root directory are these three entries.

This commit supports searching for these 3 entries and dumping them
only if they exist.

Signed-off-by: Yuezhang Mo <Yuezhang.Mo@sony.com>
Reviewed-by: Hyunchul Lee <hyc.lee@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>